[HttpClient] Refactor UrlBuilder to be an interface
